refactor: move block-file I/O helpers into node/blockstorage
Continues the incremental main.h/main.cpp decomposition (#3030). Moves the
block-file I/O helpers out of the main.{h,cpp} monolith into the existing
node/blockstorage.{h,cpp} translation unit, alongside WriteBlockToDisk /
ReadBlockFromDisk which already live there and already call them:

  CheckDiskSpace, OpenBlockFile, AppendBlockFile, LoadExternalBlockFile
  (+ the file-local BlockFilePath helper)

Their two file-local statics move with them (each has a single user):
nMinDiskSpace -> CheckDiskSpace, nCurrentBlockFile -> AppendBlockFile.

main.h keeps the consumers working by #include "node/blockstorage.h", so the
~79 includers that call these functions resolve transitively with no edits.
The declarations are removed from main.h.

Pure code move, no behavior change. Two incidental fixes that the move exposes:
- blockstorage.cpp has no `using namespace std`, so the two unqualified
  `string` locals in the moved bodies are qualified to `std::string`.
- blockstorage.cpp now sees its own header (via main.h), so the redundant
  default arguments on the two ReadBlockFromDisk *definitions* are dropped --
  default args already live on the blockstorage.h declarations, and giving
  them in both places is ill-formed (-fpermissive). Behavior is unchanged.
