+++ src/node/blockstorage.h
@@ -8,6 +8,10 @@
 
 #include "protocol.h"
 
+#include <cstddef>
+#include <cstdint>
+#include <cstdio>
+
 class CBlock;
 class CBlockIndex;
 
@@ -19,6 +23,12 @@ bool WriteBlockToDisk(const CBlock& block, unsigned int& nFileRet, unsigned int&
 
 bool ReadBlockFromDisk(CBlock& block, unsigned int nFile, unsigned int nBlockPos, const Consensus::Params& params, bool fReadTransactions=true);
 bool ReadBlockFromDisk(CBlock& block, const CBlockIndex* pindex, const Consensus::Params& params, bool fReadTransactions=true);
+
+bool CheckDiskSpace(uint64_t nAdditionalBytes=0);
+FILE* OpenBlockFile(unsigned int nFile, unsigned int nBlockPos, const char* pszMode="rb");
+FILE* AppendBlockFile(unsigned int& nFileRet);
+bool LoadExternalBlockFile(FILE* fileIn, size_t file_size = 0,
+                           unsigned int percent_start = 0, unsigned int percent_end = 100);
 
 
 #endif // BITCOIN_NODE_BLOCKSTORAGE_H
